DjangoServer/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
GoogleNews==1.6.8
matplotlib==3.7.0
numpy==1.23.5
sklearn==0.0.post1
sqlparse==0.4.3
tzdata==2022.5
pandas==1.5.3
plotly==5.13.0
Pillow==9.4.0
vaderSentiment==3.3.2
yfinance==0.2.27

